+++ src/lib.rs
@@ -16,6 +16,7 @@ mod cli_cmds;
 pub mod config;
 pub mod dataflow;
 pub mod errors;
+pub mod export;
 pub mod extract;
 pub mod graph;
 pub mod hook;